refactor: move `WinnerModal` to `modules/board/components`
