+++ src/modules/board/components/WinnerModal.jsx
